Updated KPI cards on site-wide stats overview
ref https://linear.app/ghost/issue/PROD-1893/wire-up-dashboard-kpis

- The KPI data on stats-wide analytics overview was so far static.
